refactor(tests): features -> contracts
remove contracts that do not make much sense,
keep very incomplete stdlib tests
