+++ .eslintignore
@@ -3,7 +3,7 @@ dist/
 coverage/
 **/output/
 ./src/test/bugs/output/
-./src/test/features/output/
+./src/test/contracts/output/
 src/func/funcfiftlib.js
 src/grammar/grammar.ohm*.ts
 src/grammar/grammar.ohm*.js
