+++ cspell.json
@@ -82,8 +82,8 @@
     "src/test/feature-intrinsics.spec.ts",
     "src/test/feature-optionals.spec.ts",
     "src/test/feature-strings.spec.ts",
-    "src/test/features/intrinsics.tact",
-    "src/test/features/strings.tact",
+    "src/test/contracts/intrinsics.tact",
+    "src/test/contracts/strings.tact",
     "src/test/compilation-fail/fail-const-eval.spec.ts",
     "stdlib/stdlib.fc"
   ]
