+++ src/test/contracts/implicit-init.tact
