+++ src/test/contracts/integer-literals.tact
