+++ src/test/contracts/random.tact
