+++ src/test/contracts/ternary.tact
