+++ src/test/feature-constants.spec.ts
@@ -1,7 +1,7 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
-import { ConstantTester } from "./features/output/constants_ConstantTester";
+import { ConstantTester } from "./contracts/output/constants_ConstantTester";
 
 describe("feature-constants", () => {
     beforeEach(() => {
