+++ src/test/feature-deep.spec.ts
@@ -1,9 +1,9 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
-import { A } from "./features/output/deep_A";
-import { B } from "./features/output/deep_B";
-import { C } from "./features/output/deep_C";
+import { A } from "./contracts/output/deep_A";
+import { B } from "./contracts/output/deep_B";
+import { C } from "./contracts/output/deep_C";
 
 describe("feature-random", () => {
     beforeEach(() => {
