+++ src/test/feature-integer-literals.spec.ts
@@ -1,7 +1,7 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
-import { IntegerLiteralsTester } from "./features/output/integer-literals_IntegerLiteralsTester";
+import { IntegerLiteralsTester } from "./contracts/output/integer-literals_IntegerLiteralsTester";
 
 describe("feature-integer-literals", () => {
     beforeEach(() => {
