+++ src/test/feature-map-traverse.spec.ts
@@ -1,7 +1,7 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
-import { MapTraverseTestContract } from "./features/output/map-traverse_MapTraverseTestContract";
+import { MapTraverseTestContract } from "./contracts/output/map-traverse_MapTraverseTestContract";
 
 describe("feature-map-traversal", () => {
     beforeEach(() => {
