+++ src/test/feature-masterchain.spec.ts
@@ -1,8 +1,8 @@
 import { Address, beginCell, toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
-import { MasterchainTester } from "./features/output/masterchain_MasterchainTester";
-import { MasterchainTester as EnabledTester } from "./features/output/masterchain-allow_MasterchainTester";
+import { MasterchainTester } from "./contracts/output/masterchain_MasterchainTester";
+import { MasterchainTester as EnabledTester } from "./contracts/output/masterchain-allow_MasterchainTester";
 
 describe("feature-masterchain", () => {
     beforeEach(() => {
