+++ src/test/feature-math.spec.ts
@@ -1,7 +1,7 @@
 import { beginCell, Dictionary, toNano } from "@ton/core";
 import { ContractSystem, randomAddress } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
-import { MathTester } from "./features/output/math_MathTester";
+import { MathTester } from "./contracts/output/math_MathTester";
 
 describe("feature-math", () => {
     beforeEach(() => {
