+++ src/test/feature-serialization.spec.ts
@@ -1,8 +1,8 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
-import { SerializationTester2 } from "./features/output/serialization-2_SerializationTester2";
-import { SerializationTester } from "./features/output/serialization_SerializationTester";
+import { SerializationTester2 } from "./contracts/output/serialization-2_SerializationTester2";
+import { SerializationTester } from "./contracts/output/serialization_SerializationTester";
 
 describe("feature-serialization", () => {
     beforeEach(() => {
