+++ src/test/feature-ternary.spec.ts
@@ -1,7 +1,7 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
-import { TernaryTester } from "./features/output/ternary_TernaryTester";
+import { TernaryTester } from "./contracts/output/ternary_TernaryTester";
 
 describe("feature-ternary", () => {
     beforeEach(() => {
