+++ src/test/feature-try-catch.spec.ts
@@ -1,7 +1,7 @@
 import { toNano } from "@ton/core";
 import { ContractSystem } from "@tact-lang/emulator";
 import { __DANGER_resetNodeId } from "../grammar/ast";
-import { TryCatchTester } from "./features/output/try-catch_TryCatchTester";
+import { TryCatchTester } from "./contracts/output/try-catch_TryCatchTester";
 
 describe("feature-ternary", () => {
     beforeEach(() => {
