+++ tact.config.json
@@ -81,18 +81,18 @@
     },
     {
       "name": "maps",
-      "path": "./src/test/features/maps.tact",
-      "output": "./src/test/features/output"
+      "path": "./src/test/contracts/maps.tact",
+      "output": "./src/test/contracts/output"
     },
     {
       "name": "map-traverse",
-      "path": "./src/test/features/map-traverse.tact",
-      "output": "./src/test/features/output"
+      "path": "./src/test/contracts/map-traverse.tact",
+      "output": "./src/test/contracts/output"
     },
     {
       "name": "optionals",
-      "path": "./src/test/features/optionals.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/optionals.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true,
         "masterchain": true
@@ -100,141 +100,141 @@
     },
     {
       "name": "serialization",
-      "path": "./src/test/features/serialization.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/serialization.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "serialization-2",
-      "path": "./src/test/features/serialization-2.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/serialization-2.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "serialization-3",
-      "path": "./src/test/features/serialization-3.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/serialization-3.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "strings",
-      "path": "./src/test/features/strings.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/strings.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "constants",
-      "path": "./src/test/features/constants.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/constants.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "math",
-      "path": "./src/test/features/math.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/math.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "dns",
-      "path": "./src/test/features/dns.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/dns.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "integer-literals",
-      "path": "./src/test/features/integer-literals.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/integer-literals.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "random",
-      "path": "./src/test/features/random.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/random.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "ordering",
-      "path": "./src/test/features/ordering.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/ordering.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "deep",
-      "path": "./src/test/features/deep-sequence.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/deep-sequence.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "bounced-routing",
-      "path": "./src/test/features/bounced-routing.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/bounced-routing.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "debug",
-      "path": "./src/test/features/debug.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/debug.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "send",
-      "path": "./src/test/features/send.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/send.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "intrinsics",
-      "path": "./src/test/features/intrinsics.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/intrinsics.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "masterchain",
-      "path": "./src/test/features/masterchain.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/masterchain.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "try-catch",
-      "path": "./src/test/features/try-catch.tact",
-      "output": "./src/test/features/output"
+      "path": "./src/test/contracts/try-catch.tact",
+      "output": "./src/test/contracts/output"
     },
     {
       "name": "masterchain-allow",
-      "path": "./src/test/features/masterchain.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/masterchain.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true,
         "masterchain": true
@@ -242,32 +242,32 @@
     },
     {
       "name": "address",
-      "path": "./src/test/features/address.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/address.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "ternary",
-      "path": "./src/test/features/ternary.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/ternary.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "implicit-init",
-      "path": "./src/test/features/implicit-init.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/implicit-init.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
     },
     {
       "name": "recursion",
-      "path": "./src/test/features/recursion.tact",
-      "output": "./src/test/features/output",
+      "path": "./src/test/contracts/recursion.tact",
+      "output": "./src/test/contracts/output",
       "options": {
         "debug": true
       }
@@ -294,6 +294,11 @@
       "options": {
         "debug": true
       }
+    },
+    {
+      "name": "stdlib",
+      "path": "./src/test/contracts/stdlib.tact",
+      "output": "./src/test/contracts/output"
     }
   ]
 }
